+++ src/main/java/com/ssrprojects/ultimatechatapp/config/utility/ThymeleafTemplateConfig.java
@@ -1,4 +1,4 @@
-package com.ssrprojects.ultimatechatapp.config;
+package com.ssrprojects.ultimatechatapp.config.utility;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
