+++ src/main/java/com/ssrprojects/ultimatechatapp/entity/enums/ProfileStatus.java
@@ -1,8 +1,6 @@
 package com.ssrprojects.ultimatechatapp.entity.enums;
 
 public enum ProfileStatus {
-
     PUBLIC,
     PRIVATE,
-
 }
